VISTA-249 attach documents in generated tenants

+++ vista-shared/vista-domain/src/main/java/com/propertyvista/domain/media/ApplicationDocument.java
@@ -40,6 +40,6 @@ public interface ApplicationDocument extends IEntity, IBoundToApplication {
 
     IPrimitive<String> filename();
 
-    IPrimitive<Long> fileSize();
+    IPrimitive<Integer> fileSize();
 
 }
